Tidy up, probably break some things
