+++ configuration/test_Config.py
@@ -28,15 +28,3 @@ class Test_Config(TestCase):
         Config().configfile_path = tempConfigFile
         Config().__update_config_from_file__()
         self.assertEqual('/data/path/here', Config().database_dir)
-
-    def test_config_setting_generated_property(self):
-        tempConfigFile = os.path.join( 'testdata', 'testsettings.yml')
-
-        with open(tempConfigFile, 'w') as f:
-            f.write("YAPOConfig:\n")
-            f.write("  current_setting_version: 3\n")
-            f.write("  yapo_path: somepath\n")
-        Config().configfile_path = tempConfigFile
-        Config().__update_config_from_file__()
-        self.assertEqual('somepath', Config().yapo_path)
-        self.assertEqual('todo', Config().site_path)
